feat: update the search bar
